server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
arrow==0.10.0
click==6.7
flask-cors==3.0.3
flask==0.12.2
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
markupsafe==1.0           # via jinja2
neo4j-driver==1.5.0
python-dateutil==2.6.1    # via arrow
six==1.11.0               # via flask-cors, python-dateutil
werkzeug==0.12.2          # via flask
